pyproject.toml
[project]
name = "arcfusion-rag-system"
version = "0.1.0"
description = "Multi-agent RAG system for PDF Q&A with web search, recursive retrieval, and autonomous decision-making"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "openai>=1.12.0",
    "pinecone>=5.0.0",
    "langchain>=0.1.0",
    "langgraph>=0.0.20",
    "langchain-openai>=0.0.5",
    "tavily-python>=0.2.0",
    "pymupdf>=1.23.0",
    "FlagEmbedding>=1.2.0",
    "tiktoken>=0.5.0",
    "python-multipart>=0.0.6",
    "httpx>=0.26.0",
    "loguru>=0.7.2",
    "streamlit>=1.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "black>=24.0.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
